\name{roll-package}
\alias{roll-package}
\alias{roll}
\docType{package}
\title{
  \packageTitle{roll}
}
\description{
  \packageDescription{roll}
}
\details{
  \code{roll} is a package that provides fast and efficient computation of rolling and expanding statistics for time-series data.
}
\author{
  \packageAuthor{roll}
}
\references{
  Hyndman, R.J. and Fan, Y. (1996). "Sample Quantiles in Statistical Packages." \emph{The American Statistician}, 50(4), 361-365.
  
  Welford, B.P. (1962). "Note on a Method for Calculating Corrected Sums of Squares and Products." \emph{Technometrics}, 4(3), 419-420.

  West, D.H.D. (1979). "Updating Mean and Variance Estimates: An Improved Method." \emph{Communications of the ACM}, 22(9), 532-535.
}