% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll_any}
\alias{roll_any}
\title{Rolling Any}
\usage{
roll_any(x, width, online = TRUE)
}
\arguments{
\item{x}{logical matrix or xts object. Rows are observations and columns are variables.}

\item{width}{integer. Window size.}

\item{online}{logical. Process observations using an online algorithm.}
}
\value{
An object of the same class and dimension as \code{x} with the rolling any
}
\description{
A function for computing rolling any of time-series data.
}
\examples{
n_vars <- 3
n_obs <- 15
x <- matrix(rnorm(n_obs * n_vars), nrow = n_obs, ncol = n_vars)

# rolling any
result <- roll_any(x < 0, 5)

}
