% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll_vif}
\alias{roll_vif}
\title{Rolling Variance Inflation Factors}
\usage{
roll_vif(x, width, comps = 1:(ncol(x) - 1), weights = rep(1, width),
  center = TRUE, scale = FALSE, min_obs = width, complete_obs = TRUE,
  na_restore = FALSE, parallel_for = c("rows", "cols"))
}
\arguments{
\item{x}{matrix or xts object. Rows are observations and columns are the independent variables.}

\item{width}{integer. Window size.}

\item{comps}{integer vector. Select a subset of principal components.}

\item{weights}{vector. Weights for each observation within a window.}

\item{center}{logical. If \code{TRUE} then the weighted mean of each variable is used,
if \code{FALSE} then zero is used.}

\item{scale}{logical. If \code{TRUE} then the weighted standard deviation of each variable is used,
if \code{FALSE} then no scaling is done.}

\item{min_obs}{integer. Minimum number of observations required to have a value within a window, 
otherwise result is NA.}

\item{complete_obs}{logical. If \code{TRUE} then rows containing any missing values are removed,
if \code{FALSE} then pairwise is used.}

\item{na_restore}{logical. Should missing values be restored?}

\item{parallel_for}{character. Executes a "for" loop in which iterations run in parallel by
\code{rows} or \code{cols}.}
}
\value{
An object of the same class and dimension as \code{x} with the rolling variance inflation factors.
}
\description{
A parallel function for computing rolling variance inflation factors of time-series data.
}
\note{
If users are already taking advantage of parallelism using multithreaded BLAS/LAPACK
libraries, then limit the number of cores in the RcppParallel package to one with the
\code{\link[RcppParallel]{setThreadOptions}} function.
}
\examples{
n_vars <- 10
n_obs <- 1000
x <- matrix(rnorm(n_obs * n_vars), nrow = n_obs, ncol = n_vars)

# 252-day rolling variance inflation factors (lm)
result <- roll_vif(x, 252)

# 252-day rolling variance inflation factors (pcr)
result <- roll_vif(x, 252, comps = 1)

# Equivalent to 'na.rm = TRUE'
result <- roll_vif(x, 252, min_obs = 1)

# Expanding window
result <- roll_vif(x, n_obs, min_obs = 1)

# Exponential decay
weights <- 0.9 ^ (251:0)
result <- roll_vif(x, 252, weights = weights, min_obs = 1)
}
\seealso{
\code{\link[RcppParallel]{setThreadOptions}} for thread options via RcppParallel.
}

