% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll_median}
\alias{roll_median}
\title{Rolling Medians}
\usage{
roll_median(x, width, weights = rep(1, width), min_obs = width,
  complete_obs = FALSE, na_restore = FALSE, online = FALSE)
}
\arguments{
\item{x}{vector or matrix. Rows are observations and columns are variables.}

\item{width}{integer. Window size.}

\item{weights}{vector. Weights for each observation within a window.}

\item{min_obs}{integer. Minimum number of observations required to have a value within a window,
otherwise result is \code{NA}.}

\item{complete_obs}{logical. If \code{TRUE} then rows containing any missing values are removed,
if \code{FALSE} then each value is used.}

\item{na_restore}{logical. Should missing values be restored?}

\item{online}{logical. Process observations using an online algorithm.}
}
\value{
An object of the same class and dimension as \code{x} with the rolling and expanding
medians.
}
\description{
A function for computing the rolling and expanding medians of time-series data.
}
\examples{
n <- 15
x <- rnorm(n)
weights <- 0.9 ^ (n:1)

# rolling medians with complete windows
roll_median(x, width = 5)

# rolling medians with partial windows
roll_median(x, width = 5, min_obs = 1)

# expanding medians with partial windows
roll_median(x, width = n, min_obs = 1)

# expanding medians with partial windows and weights
roll_median(x, width = n, min_obs = 1, weights = weights)
}
